storage/reports: move a test
Move the unit test for the constraints conformance report to
constraint_report_test.go from reporter_test.go, since it's specific to
one type of report.

Release note: None
